・タイトル、セレクトの中途制作物のマージ完了 ・ゲームクリア後に緑化がリセットされない問題を解消 ・緑化サイズ、速度の調整 ・背景でスクロールするエフェクトの仮実装(動きのみ)

+++ CaseStudy/CaseStudy/res/cpp/Object/Jack.cpp
@@ -48,9 +48,9 @@ void CJack::Init(D3DXVECTOR3 pos,D3DXVECTOR3 dir)
 	m_angle = AngleOf2Vector(dir,D3DXVECTOR3(0,1,0));
 
 	if(dir.x > 0)
-		RotateZ(-m_angle);
+		RotateZ(-(float)m_angle);
 	if(dir.x < 0)
-		RotateZ(m_angle);
+		RotateZ((float)m_angle);
 
 	// アニメーション初期化
 	StartAnimation();
